+++ broadcasterctl/BroadcasterAdminController.m
@@ -1,9 +1,9 @@
 /*
  *
  * @APPLE_LICENSE_HEADER_START@
- * 
- * Copyright (c) 1999-2003 Apple Computer, Inc.  All Rights Reserved.
- * 
+ *
+ * Copyright (c) 1999-2008 Apple Inc.  All Rights Reserved.
+ *
  * This file contains Original Code and/or Modifications of Original Code
  * as defined in and that are subject to the Apple Public Source License
  * Version 2.0 (the 'License'). You may not use this file except in
@@ -101,7 +101,7 @@
 - (NSString *)resultForArgument:(NSString *)argName foundError:(BOOL *)err
 {
     NSArray *args = [[NSProcessInfo processInfo] arguments];
-    int argLoc;
+    unsigned long argLoc;
     NSString *result;
 
     argLoc = [args indexOfObject:argName];
